+++ spec/unit/installer/xcode/pods_project_generator/pod_target_integrator_spec.rb
@@ -14,7 +14,7 @@ module Pod
               @watermelon_spec = fixture_spec('watermelon-lib/WatermelonLib.podspec')
               @watermelon_pod_target = fixture_pod_target_with_specs([@watermelon_spec,
                                                                       *@watermelon_spec.recursive_subspecs],
-                                                                     BuildType.dynamic_framework, {}, [], Platform.ios,
+                                                                     BuildType.static_library, {}, [], Platform.ios,
                                                                      [@target_definition])
 
               @coconut_spec = fixture_spec('coconut-lib/CoconutLib.podspec')
@@ -42,6 +42,45 @@ module Pod
                 @test_native_target.build_phases[0].shell_script.should == "\"${PODS_ROOT}/Target Support Files/CoconutLib/CoconutLib-Unit-Tests-frameworks.sh\"\n"
               end
 
+              it '*does* integrate test native targets with resources script phase if built statically' do
+                resource_paths = ['${PODS_CONFIGURATION_BUILD_DIR}/TestResourceBundle.bundle']
+                @coconut_pod_target.stubs(:resource_paths).returns('CoconutLib' => resource_paths)
+                @coconut_pod_target.stubs(:build_type).returns(BuildType.static_library)
+                PodTargetIntegrator.new(@coconut_target_installation_result).integrate!
+                @test_native_target.build_phases.count.should == 1
+                @test_native_target.build_phases.map(&:display_name).should == [
+                  '[CP] Copy Pods Resources',
+                ]
+                @test_native_target.build_phases[0].shell_script.should == "\"${PODS_ROOT}/Target Support Files/CoconutLib/CoconutLib-Unit-Tests-resources.sh\"\n"
+                @test_native_target.build_phases[0].input_paths.should == [
+                  '${PODS_ROOT}/Target Support Files/CoconutLib/CoconutLib-Unit-Tests-resources.sh',
+                  '${PODS_CONFIGURATION_BUILD_DIR}/TestResourceBundle.bundle',
+                ]
+                @test_native_target.build_phases[0].output_paths.should == [
+                  '${TARGET_BUILD_DIR}/${UNLOCALIZED_RESOURCES_FOLDER_PATH}/TestResourceBundle.bundle',
+                ]
+              end
+
+              it 'does not integrates test native targets with resources script phase if built dynamically' do
+                framework_paths = [Pod::Xcode::FrameworkPaths.new('${PODS_ROOT}/Vendored/Vendored.framework')]
+                resource_paths = ['${PODS_CONFIGURATION_BUILD_DIR}/TestResourceBundle.bundle']
+                @coconut_pod_target.stubs(:framework_paths).returns('CoconutLib' => framework_paths)
+                @coconut_pod_target.stubs(:resource_paths).returns('CoconutLib' => resource_paths)
+                PodTargetIntegrator.new(@coconut_target_installation_result).integrate!
+                @test_native_target.build_phases.count.should == 1
+                @test_native_target.build_phases.map(&:display_name).should == [
+                  '[CP] Embed Pods Frameworks',
+                ]
+                @test_native_target.build_phases[0].shell_script.should == "\"${PODS_ROOT}/Target Support Files/CoconutLib/CoconutLib-Unit-Tests-frameworks.sh\"\n"
+                @test_native_target.build_phases[0].input_paths.should == [
+                  '${PODS_ROOT}/Target Support Files/CoconutLib/CoconutLib-Unit-Tests-frameworks.sh',
+                  '${PODS_ROOT}/Vendored/Vendored.framework',
+                ]
+                @test_native_target.build_phases[0].output_paths.should == [
+                  '${TARGET_BUILD_DIR}/${FRAMEWORKS_FOLDER_PATH}/Vendored.framework',
+                ]
+              end
+
               it 'clears input and output paths from script phase if it exceeds limit' do
                 # The paths represented here will be 501 for input paths and 501 for output paths
                 # which will exceed the limit.
@@ -49,13 +88,13 @@ module Pod
                   "${PODS_CONFIGURATION_BUILD_DIR}/DebugLib/DebugLibPng#{i}.png"
                 end
                 @coconut_pod_target.stubs(:resource_paths).returns('CoconutLib' => resource_paths)
+                @coconut_pod_target.stubs(:build_type).returns(BuildType.static_library)
                 PodTargetIntegrator.new(@coconut_target_installation_result).integrate!
                 @test_native_target.build_phases.map(&:display_name).should == [
-                  '[CP] Embed Pods Frameworks',
                   '[CP] Copy Pods Resources',
                 ]
-                @test_native_target.build_phases[1].input_paths.should == []
-                @test_native_target.build_phases[1].output_paths.should == []
+                @test_native_target.build_phases[0].input_paths.should == []
+                @test_native_target.build_phases[0].output_paths.should == []
               end
 
               it 'integrates test native targets with frameworks and resources script phase input and output paths' do
@@ -63,6 +102,7 @@ module Pod
                 resource_paths = ['${PODS_CONFIGURATION_BUILD_DIR}/TestResourceBundle.bundle']
                 @coconut_pod_target.stubs(:framework_paths).returns('CoconutLib' => framework_paths)
                 @coconut_pod_target.stubs(:resource_paths).returns('CoconutLib' => resource_paths)
+                @coconut_pod_target.stubs(:build_type).returns(BuildType.static_library)
                 PodTargetIntegrator.new(@coconut_target_installation_result, :use_input_output_paths => true).integrate!
                 @test_native_target.build_phases.count.should == 2
                 @test_native_target.build_phases.map(&:display_name).should == [
@@ -91,6 +131,7 @@ module Pod
                 resource_paths = ['${PODS_CONFIGURATION_BUILD_DIR}/TestResourceBundle.bundle']
                 @coconut_pod_target.stubs(:framework_paths).returns('CoconutLib' => framework_paths)
                 @coconut_pod_target.stubs(:resource_paths).returns('CoconutLib' => resource_paths)
+                @coconut_pod_target.stubs(:build_type).returns(BuildType.static_library)
                 PodTargetIntegrator.new(@coconut_target_installation_result, :use_input_output_paths => true).integrate!
                 @test_native_target.build_phases.count.should == 2
                 @test_native_target.build_phases.map(&:display_name).should == [
@@ -118,6 +159,7 @@ module Pod
                 resource_paths = ['${PODS_CONFIGURATION_BUILD_DIR}/TestResourceBundle.bundle']
                 @coconut_pod_target.stubs(:framework_paths).returns('CoconutLib' => framework_paths)
                 @coconut_pod_target.stubs(:resource_paths).returns('CoconutLib' => resource_paths)
+                @coconut_pod_target.stubs(:build_type).returns(BuildType.static_library)
                 PodTargetIntegrator.new(@coconut_target_installation_result, :use_input_output_paths => true).integrate!
                 @test_native_target.build_phases.count.should == 2
                 @test_native_target.build_phases.map(&:display_name).should == [
@@ -228,6 +270,7 @@ module Pod
                 resource_paths = ['${PODS_CONFIGURATION_BUILD_DIR}/TestResourceBundle.bundle']
                 @coconut_pod_target.stubs(:framework_paths).returns('CoconutLib' => framework_paths)
                 @coconut_pod_target.stubs(:resource_paths).returns('CoconutLib' => resource_paths)
+                @coconut_pod_target.stubs(:build_type).returns(BuildType.static_library)
                 PodTargetIntegrator.new(@coconut_target_installation_result, :use_input_output_paths => false).integrate!
                 PodTargetIntegrator.new(@coconut_target_installation_result, :use_input_output_paths => false).integrate!
                 @test_native_target.build_phases.count.should == 2